+++ src/main/java/nl/jusx/pycharm/lineprofiler/render/LineProfileInlayRenderer.java
@@ -9,6 +9,7 @@ import com.intellij.openapi.editor.colors.TextAttributesKey;
 import com.intellij.openapi.editor.markup.TextAttributes;
 import nl.jusx.pycharm.lineprofiler.profile.LineProfile;
 import nl.jusx.pycharm.lineprofiler.service.ColorMapService;
+import nl.jusx.pycharm.lineprofiler.settings.SettingsState;
 import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
@@ -22,7 +23,6 @@ import static nl.jusx.pycharm.lineprofiler.render.InlayRendererUtils.*;
  * Inspired by {@link com.intellij.xdebugger.impl.inline.InlineDebugRenderer}
  */
 public class LineProfileInlayRenderer implements EditorCustomElementRenderer {
-    private static final int MAX_TABLE_ALIGNMENT_IN_CHARS = 75;
     private static final int RESULT_TABLE_STRING_MARGIN_BLOCKS = 3;
 
     private final LineProfile lineProfile;
@@ -58,7 +58,7 @@ public class LineProfileInlayRenderer implements EditorCustomElementRenderer {
     private void paintTableAligned(@NotNull Editor editor, @NotNull Graphics g, @NotNull Point renderAnchor) {
         // align table
         renderAnchor.x += 3 * margin;
-        if (renderAnchor.x < getFontMetrics(editor).charWidth(' ') * MAX_TABLE_ALIGNMENT_IN_CHARS) {
+        if (renderAnchor.x < getFontMetrics(editor).charWidth(' ') * SettingsState.getInstance().getTableAlignmentMaxColumns()) {
             renderAnchor.x = tableAlignment.align(renderAnchor.x, String.valueOf(lineProfile.getLineNrFromZero()));
         }