+++ src/main/java/nl/jusx/pycharm/lineprofiler/settings/SettingsComponent.java
@@ -1,5 +1,6 @@
 package nl.jusx.pycharm.lineprofiler.settings;
 
+import com.intellij.ui.JBIntSpinner;
 import com.intellij.ui.components.JBLabel;
 import com.intellij.util.ui.FormBuilder;
 
@@ -14,19 +15,11 @@ import javax.swing.*;
 public class SettingsComponent {
     private final JPanel myMainPanel;
 
-//    private final ButtonGroup defaultTimeFractionCalculation = new ButtonGroup();
-//    private final JBRadioButton functionTotalButton = new JBRadioButton("Total function times", true);
-//    private final JBRadioButton profileTotalButton = new JBRadioButton("Total profile time");
+    private final JBIntSpinner myMaxTableAlignment = new JBIntSpinner(120, 0, 300);
 
     public SettingsComponent() {
-//        defaultTimeFractionCalculation.add(profileTotalButton);
-//        defaultTimeFractionCalculation.add(functionTotalButton);
-
         myMainPanel = FormBuilder.createFormBuilder()
-                .addComponent(new JBLabel("No settings exist for 'Line Profiler Plugin'"))
-//                .addComponent(new JBLabel("As a default, visualizations should express percentages of"), 1)
-//                .addComponent(functionTotalButton, 1)
-//                .addComponent(profileTotalButton, 1)
+                .addLabeledComponent(new JBLabel("Max table alignment at column: "), myMaxTableAlignment, 1, false)
                 .addComponentFillVertically(new JPanel(), 0)
                 .getPanel();
     }
@@ -35,23 +28,15 @@ public class SettingsComponent {
         return myMainPanel;
     }
 
-//    public JComponent getPreferredFocusedComponent() {
-//        return profileTotalButton;
-//    }
-//
-//    public void setDefaultTimeFractionCalculation(TimeFractionCalculation timeFractionCalculation) {
-//        if (timeFractionCalculation == TimeFractionCalculation.PROFILE_TOTAL) {
-//            defaultTimeFractionCalculation.setSelected(profileTotalButton.getModel(), true);
-//        } else {
-//            defaultTimeFractionCalculation.setSelected(functionTotalButton.getModel(), true);
-//        }
-//    }
-//
-//    public TimeFractionCalculation getDefaultTimeFractionCalculation() {
-//        if (defaultTimeFractionCalculation.getSelection() == profileTotalButton.getModel()) {
-//            return TimeFractionCalculation.PROFILE_TOTAL;
-//        } else {
-//            return TimeFractionCalculation.FUNCTION_TOTAL;
-//        }
-//    }
+    public JComponent getPreferredFocusedComponent() {
+        return myMaxTableAlignment;
+    }
+
+    public void setMaxTableAlignment(int maxTableAlignment) {
+        myMaxTableAlignment.setValue(maxTableAlignment);
+    }
+
+    public int getMaxTableAlignment() {
+        return (int) myMaxTableAlignment.getValue();
+    }
 }