+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Made the max table alignment column configurable
+
 ### Fixed
 - Use longs for 'time' and 'hits' instead of ints to prevent NumberFormatExceptions